feat: challengelist 컴포넌트 내 api 호출 함수 정리 및 route handler로 적절히 분배 작업 (진행 중, 1개 완료)
