+++ src/api/supabase/challenge.ts
@@ -1,11 +1,13 @@
+import { createClient } from "@/supabase/client"
 import axios from "axios"
 
-import { Database } from "../../../types/supabase"
-
-// 테이블 상에 존재하는 데이터 타입
-type ChallengeType = Database["public"]["Tables"]["challenge"]["Row"]
-type MilestoneType = Database["public"]["Tables"]["milestone"]["Row"]
-export type RoutineType = Database["public"]["Tables"]["routine"]["Row"]
+import {
+  ChallengeType,
+  MilestoneType,
+  RoutineType,
+  StructuredChallengeType,
+  StructuredMilestoneType,
+} from "../../../types/supabase.type"
 
 // 필요없는 타입 제외한 milestone 타입
 type MilestoneDefaultType = Pick<
@@ -44,6 +46,20 @@ export interface POSTchallengeArgumentProps {
   milestone: (MilestoneRequiredType & MilestonePartialType)[]
   routine: Pick<RoutineType, "content" | "milestone_id">[][]
 }
+
+// 구조화 챌린지 가져오기 함수 인자 타입
+export interface GetstructuredChallengeArgumentProps {
+  "user-id": string
+}
+
+// 챌린지 디테일 업데이트 함수 인자 타입
+export interface PUTchallengeArgumentProps {
+  "challenge-id": string
+  milestoneIds: ChallengeType["id"][]
+  milestone: (MilestoneRequiredType & RemainingType)[]
+  routine: Pick<RoutineType, "content" | "milestone_id">[][]
+}
+
 // 챌린지 생성함수
 export const POSTchallenge = async (params: POSTchallengeArgumentProps) => {
   const postResponse = await axios.post("/api/challenge", {
@@ -54,13 +70,6 @@ export const POSTchallenge = async (params: POSTchallengeArgumentProps) => {
   return postResponse
 }
 
-// 챌린지 디테일 업데이트 함수 인자 타입
-export interface PUTchallengeArgumentProps {
-  "challenge-id": string
-  milestoneIds: ChallengeType["id"][]
-  milestone: (MilestoneRequiredType & RemainingType)[]
-  routine: Pick<RoutineType, "content" | "milestone_id">[][]
-}
 // 챌린지 디테일 업데이트 함수
 export const PUTchallenge = async (params: PUTchallengeArgumentProps) => {
   const putResponse = await axios.put(
@@ -72,4 +81,11 @@ export const PUTchallenge = async (params: PUTchallengeArgumentProps) => {
     }
   )
   return putResponse
+}
+
+// 유저 ID 기반으로 구조화된 챌린지-마일스톤-루틴 데이터 일괄적으로 가져오는 함수
+export const GETstructuredChallengeData = async (userId: string) => {
+  const response = await axios.get(`/api/structured-challenge/${userId}`)
+  const getResponse: StructuredChallengeType[] = response.data
+  return getResponse
 }
