+++ src/app/(providers)/(styles)/my-page/profile/page.tsx
@@ -102,7 +102,6 @@ function UserInfoPage() {
   }
 
   const handleImageChange = (e: ChangeEvent<HTMLInputElement>) => {
-
     const file = e.target.files?.[0]
     if (file) {
       setSelectedFile(file)
