+++ types/supabase.type.ts
@@ -1,0 +1,17 @@
+import { Database } from "./supabase"
+
+export type ChallengeType = Database["public"]["Tables"]["challenge"]["Row"]
+
+export type StructuredChallengeType =
+  Database["public"]["Tables"]["challenge"]["Row"] & {
+    milestones: StructuredMilestoneType[]
+  }
+
+export type MilestoneType = Database["public"]["Tables"]["milestone"]["Row"]
+
+export type StructuredMilestoneType =
+  Database["public"]["Tables"]["milestone"]["Row"] & {
+    routines: RoutineType[]
+  }
+
+export type RoutineType = Database["public"]["Tables"]["routine"]["Row"]
